:recycle: Update the types for web3-eth
