+++ packages/web3-eth/src/rpc_methods.ts
@@ -23,8 +23,8 @@ import {
 	HexString32Bytes,
 	HexString8Bytes,
 	HexStringBytes,
-	TransactionCall,
-	TransactionWithSender,
+	TransactionCallAPI,
+	TransactionWithSenderAPI,
 	Uint,
 	Uint256,
 } from 'web3-types';
@@ -295,7 +295,7 @@ export async function sign(
  */
 export async function signTransaction(
 	requestManager: Web3RequestManager,
-	transaction: TransactionWithSender | Partial<TransactionWithSender>,
+	transaction: TransactionWithSenderAPI | Partial<TransactionWithSenderAPI>,
 ) {
 	return requestManager.send({
 		method: 'eth_signTransaction',
@@ -314,7 +314,7 @@ export async function signTransaction(
  */
 export async function sendTransaction(
 	requestManager: Web3RequestManager,
-	transaction: TransactionWithSender | Partial<TransactionWithSender>,
+	transaction: TransactionWithSenderAPI | Partial<TransactionWithSenderAPI>,
 ) {
 	return requestManager.send({
 		method: 'eth_sendTransaction',
@@ -348,7 +348,7 @@ export async function sendRawTransaction(
  */
 export async function call(
 	requestManager: Web3RequestManager,
-	transaction: TransactionCall,
+	transaction: TransactionCallAPI,
 	blockNumber: BlockNumberOrTag,
 ) {
 	// validateTransactionCall(transaction);
@@ -369,7 +369,7 @@ export async function call(
  */
 export async function estimateGas(
 	requestManager: Web3RequestManager,
-	transaction: Partial<TransactionWithSender>,
+	transaction: Partial<TransactionWithSenderAPI>,
 	blockNumber: BlockNumberOrTag,
 ) {
 	validator.validate(['blockNumberOrTag'], [blockNumber]);
