+++ packages/web3-eth/src/utils/get_transaction_gas_pricing.ts
@@ -17,12 +17,12 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { FormatType, DataFormat, format, ETH_DATA_FORMAT } from 'web3-utils';
 import { Web3Context } from 'web3-core';
-import { EthExecutionAPI, Numbers } from 'web3-types';
+import { EthExecutionAPI, Numbers, Transaction } from 'web3-types';
 import { isNullish } from 'web3-validator';
 import { Eip1559NotSupportedError, UnsupportedTransactionTypeError } from '../errors';
 // eslint-disable-next-line import/no-cycle
 import { getBlock, getGasPrice } from '../rpc_method_wrappers';
-import { InternalTransaction, Transaction } from '../types';
+import { InternalTransaction } from '../types';
 // eslint-disable-next-line import/no-cycle
 import { getTransactionType } from './transaction_builder';
