+++ packages/web3-eth/src/utils/wait_for_transaction_receipt.ts
@@ -14,11 +14,10 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { EthExecutionAPI, Bytes } from 'web3-types';
+import { EthExecutionAPI, Bytes, TransactionReceipt } from 'web3-types';
 import { Web3Context } from 'web3-core';
 import { DataFormat, isNullish } from 'web3-utils';
 
-import { TransactionReceipt } from '../types';
 // eslint-disable-next-line import/no-cycle
 import { getTransactionReceipt } from '../rpc_method_wrappers';
 import { TransactionPollingTimeoutError } from '../errors';
