+++ packages/web3-eth/test/integration/helper.ts
@@ -14,11 +14,11 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { Block } from 'web3-types';
+import { Block, TransactionInfo, TransactionReceipt } from 'web3-types';
 import { DEFAULT_RETURN_FORMAT, FMT_NUMBER } from 'web3-utils';
 import { Web3PromiEvent } from 'web3-core';
 import { AbiEventFragment } from 'web3-eth-abi';
-import { TransactionReceipt, SendTransactionEvents, TransactionInfo, Web3Eth } from '../../src';
+import { SendTransactionEvents, Web3Eth } from '../../src';
 import { BasicAbi } from '../shared_fixtures/build/Basic';
 
 type SendFewTxParams = {
