+++ packages/web3-eth/test/unit/rpc_method_wrappers/fixtures/get_logs.ts
@@ -14,9 +14,9 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { FilterResults, Filter } from 'web3-types';
+import { FilterResultsAPI, Filter } from 'web3-types';
 
-export const mockRpcResponse: FilterResults = [
+export const mockRpcResponse: FilterResultsAPI = [
 	{
 		logIndex: '0x1',
 		blockNumber: '0x1b4',
