+++ packages/web3-eth/test/unit/rpc_method_wrappers/fixtures/get_pending_transactions.ts
@@ -14,7 +14,7 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { Transaction } from '../../../../src/types';
+import { Transaction } from 'web3-types';
 
 const transaction: Transaction = {
 	from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
